requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
gtfs-realtime-bindings==1.0.0
protobuf==6.33.5
psycopg2-binary==2.9.11
python-dotenv==1.0.0
redis==4.5.4
requests==2.31.0
SQLAlchemy==2.0.47
